fix: repair build errors across Rust backend and TypeScript frontend
- Fix JSX nesting in TranscriptSettings.tsx (unclosed div)
- Add missing src/lib/transport.ts (Tauri invoke re-export)
- Fix corrupted Transcript interface in types/index.ts
- Add transient_speaker to VirtualizedTranscriptView destructuring
- Patch whisper-rs-sys for Windows bindgen bug (bundled bindings)
- Fix lifetime/ownership errors in recording_saver and transcription worker
- Fix doc comments on function parameters in whisper_engine
- Fix mismatched error types in summary/failover and llm_postprocess
- Fix f32/f64 mismatch in diarization clustering
- Fix raw string literal closing in summary/processor
- Move misplaced test functions out of impl Default block
- Add .cargo-patches/ to .gitignore

+++ frontend/src-tauri/src/diarization/offline.rs
@@ -62,7 +62,7 @@ pub async fn commit_speaker_labels(
     for (seg_id, seg_start, seg_end) in segments {
         let mid = (seg_start + seg_end) / 2.0;
         let mut best_idx = 0usize;
-        let mut best_dist = f32::MAX;
+        let mut best_dist = f64::MAX;
         for (i, w) in windows.iter().enumerate() {
             let d = (mid - (w.audio_start + w.audio_end) / 2.0).abs();
             if d < best_dist {
